Moving helpers to testHelpers directory
