+++ src/app/components/Figure/Image/index.amp.test.jsx
@@ -1,5 +1,5 @@
 import AmpImg from './index.amp';
-import snapshotTests from './helpers/snapshotTests';
+import snapshotTests from './testHelpers/snapshotTests';
 
 describe('Image - AmpImg', () => {
   snapshotTests(AmpImg);
