+++ src/app/components/Figure/Image/testHelpers/fixtureData.js
