+++ src/app/components/Figure/Image/testHelpers/snapshotTests.jsx
