+++ src/app/components/Figure/Image/testHelpers/stories.jsx
